refactor slide component
