+++ src/components/Slide/index.js
@@ -1,0 +1,3 @@
+import Slide from "./Slide";
+
+export default Slide;
